Removed comments and changed ubuntu version
There is an issue when running EPICS on ubuntu 24:
https://epics.anl.gov/tech-talk/2024/msg00761.php

Essentially ubuntu changed the default value for the following compiler
flag:
https://wiki.ubuntu.com/ToolChain/CompilerFlags#A-D_FORTIFY_SOURCE.3D3

Now this triggers when running an ioc. I changed to ubuntu 22 to get
around this but you can also change the flag value to solve it.

I also removed old code that I used to troubleshoot stuff (and I forgot
to remove it previously)